+++ packages/puer_time_travel/lib/puer_time_travel.dart
@@ -1,1 +1,3 @@
+export 'package:puer/puer.dart';
+
 export 'src/time_travel_controller.dart';