+++ samples/time_travel_example/lib/domain/counter_feature.dart
@@ -1,4 +1,3 @@
-import 'package:puer/puer.dart';
 import 'package:puer_time_travel/puer_time_travel.dart';
 
 import '../data/counter_storage.dart';